moved normalizeConnectorId to shared utils
