+++ libs/designer/src/lib/core/templates/utils/helper.ts
@@ -1,4 +1,4 @@
-import { type Template, getIntl } from '@microsoft/logic-apps-shared';
+import { type Template, getIntl, normalizeConnectorId } from '@microsoft/logic-apps-shared';
 import type { AppDispatch } from '../../../core';
 import { summaryTab } from '../../../ui/panel/templatePanel/quickViewPanel/tabs/summaryTab';
 import { workflowTab } from '../../../ui/panel/templatePanel/quickViewPanel/tabs/workflowTab';
@@ -6,7 +6,6 @@ import type { IntlShape } from 'react-intl';
 import type { FilterObject } from '@microsoft/designer-ui';
 import Fuse from 'fuse.js';
 import { validateParameterValueWithSwaggerType } from '../../../core/utils/validation';
-import { normalizeConnectorId } from '../../../core/utils/connectors/connections';
 
 export const getQuickViewTabs = (intl: IntlShape, dispatch: AppDispatch, { templateId, workflowAppName }: Template.TemplateContext) => {
   return [
