+++ libs/logic-apps-shared/src/utils/src/lib/helpers/__test__/connectors.spec.tsx
@@ -5,6 +5,7 @@ import {
   getDescriptionFromConnector,
   getDisplayNameFromConnector,
   getIconUriFromConnector,
+  normalizeConnectorId,
 } from '../index';
 import { describe, vi, beforeEach, afterEach, beforeAll, afterAll, it, test, expect } from 'vitest';
 describe('utils/src/lib/helpers/connectors', () => {
@@ -208,6 +209,25 @@ describe('utils/src/lib/helpers/connectors', () => {
 
     test('works for undefined', () => {
       expect(getDescriptionFromConnector(undefined)).toBe('');
+    });
+  });
+
+  describe('normalizeConnectorId', () => {
+    const armConnectorId = '/subscriptions/#subscription#/providers/Microsoft.Web/locations/#location#/managedApis/sql';
+    const subscriptionId = '00000000-0000-0000-0000-000000000000';
+    const location = 'eastus';
+
+    it('should replace subscriptionId and location correctly in arm connector id', async () => {
+      const expectedConnectorId = `/subscriptions/${subscriptionId}/providers/Microsoft.Web/locations/${location}/managedApis/sql`;
+      expect(normalizeConnectorId(armConnectorId, subscriptionId, location)).toEqual(expectedConnectorId);
+
+      expect(normalizeConnectorId(armConnectorId, '', '')).toEqual('/subscriptions//providers/Microsoft.Web/locations//managedApis/sql');
+    });
+
+    it('should not change connectorId when not an arm resource', async () => {
+      expect(normalizeConnectorId('', subscriptionId, location)).toEqual('');
+      expect(normalizeConnectorId('/serviceProviders/sql', subscriptionId, location)).toEqual('/serviceProviders/sql');
+      expect(normalizeConnectorId('/dataOperations', '', '')).toEqual('/dataOperations');
     });
   });
 });
